Add the ability to monitor file paths for changes
Add a file watcher to monitor open files and update the associated
zee buffer as the content changes. Event notification is done by the
`notify` crate using the recommended notification API (inotify,
fsevent, etc.) of the host platform.

Much of the watcher module code is taken from the xi-editor project
which is licensed under the Apache 2.0 license.